[refactor] #283 MyDrawCourse Domain 패키지로 이동
