+++ app/src/main/res/layout/item_storage_my_draw.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="storageItem"
-            type="com.runnect.runnect.data.dto.MyDrawCourse" />
+            type="com.runnect.runnect.domain.entity.MyDrawCourse" />
 
         <variable
             name="selected"
